refactor(isa): extract x86 stochastic candidate generation into a tested seam
Lift the x86 stochastic-search candidate-space policy out of the
~5000-line `src/isa/x86/mod.rs` backend into a dedicated
`src/isa/x86/stochastic.rs` submodule, mirroring the existing
`isa/x86/encoding.rs` seam.

The moved cluster is the generator (`X86InstructionGenerator`), the MCMC
mutator (`X86Mutator`), the shared `build_x86_instruction_by_opcode`
dispatch table, the rewritable-opcode/CMOV rules, the default register
and immediate pools, and the operand-rewrite tables. Generator and
mutator move together because they are deliberately RNG-lock-stepped
through the shared dispatch table; splitting them across files would
break that locality.

The move is behaviour-preserving: the cluster's characterization tests
move with it into `stochastic.rs`, and the public surface
(`X86InstructionGenerator`, `X86Mutator`, `default_x86_registers`,
`default_x86_immediates`) is re-exported from `isa::x86` so every call
site in the search backends is byte-identical. `mod.rs` now keeps only
the `Assembler::can_assemble` encodability prefilter it still needs.

Verification: `cargo test --lib -- --list` yields the identical 1689
leaf-test set before and after; all 1689 pass; the moved code is
byte-for-byte identical to the original; `./ci_check.sh` is green.
